requirements.txt
phidata
openai
duckduckgo-search
yfinance
groq
streamlit